app/requirements.txt
Flask==2.2.3
mysql-connector-python==8.0.28
pyjwt==2.6.0
werkzeug==2.2.3